reorganize folders into pck and experiments
