+++ src/qcm/components/encoder/pcam.py
