+++ src/qcm/components/hamiltonian/embedding.py
