+++ src/qcm/components/hamiltonian/embedding_light.py
