+++ src/qcm/components/qnn/amplitude_encoding.py
