+++ src/qcm/components/qnn/angle_encoding.py
