+++ src/qcm/components/qnn/classical.py
